Allow more then one month to be drawn in the mini calendar
It is annoying at the end of a month to not have an over view of the
next month. Add a configuration option showMiniMonthCalendarMonths to
allow specifying how many months after the current one shall be drawn.

This is backward compatible with current configurations as the boolean
flag showMiniMonthCalendar is kept to gate if a mini calendar should be
drawn at all. If in future the configuration syntax changes this boolean
flag could be dropped in favor of disabling the min i calendar by
specifying 0 months to draw in the new showMiniMonthCalendarMonths
option.